rename rtclienttest
