+++ test/src/realtime/test_realtime.cpp
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "realtime/RtClientTestBase.h"
+#include "realtime/RtClientTest.h"
 #include <thread>
 
 namespace Nakama {
