+++ test/src/realtime/test_rpc.cpp
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "realtime/RtClientTestBase.h"
+#include "realtime/RtClientTest.h"
 #include "TaskExecutor.h"
 #include "test_serverConfig.h"
