+++ test/src/realtime/test_tournament.cpp
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "realtime/RtClientTestBase.h"
+#include "realtime/RtClientTest.h"
 #include "nakama-cpp/NUtils.h"
 #define RAPIDJSON_HAS_STDSTRING 1
 #include "rapidjson/document.h"
